requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
python-docx==1.1.0
redis==5.0.1
pydantic==2.5.0
httpx==0.25.2
edge-tts==6.1.9
pygame==2.5.2
asyncio-mqtt==0.16.1
jinja2==3.1.2